moving the common methods to an abstract class WIP
